Clean up chat submit actions
